Fixed getting labels and get_commit error handling
An error getting the most recent commit in get_labels was causing images to be incorrectly labelled as not jot generated, which resulted in create_local_image being unable to find the image it correctly created. I modified get_commit slightly so that it is able to properly catch errors so that images can be labelled correctly. I also modified get_labels(image_inspect::AbstractDict{String, Any}) because it was pulling labels from the wrong key.